Return error code when composing PDU fails When the PduComposer fails to make the message body, a successful result code was still returned.
Change-Id: I8a677afeeae8811c3b0fae61209c5348df1e39d9

Increase PDU parts buffer size for performance reasons
Changed the size of the temporary buffer used when storing a PDU part
to 8192 bytes instead of the previous 256 bytes. This greatly
decreases the time needed to store relatively large PDU parts. The
times to store PDU parts were so long that we frequently ended up with
an ANR. This change resulted in a total time usage of ~1000 ms instead
of ~10000 ms for ~500 kB worth of data.

Change-Id: Ia02cb28e4fd9dfe3aaa1fa30ff37659951cbed93