blueti: add corresponding remove_match for GATT
With introduction of blueti, a new dbus match was added to filter events out
of blueZ's org.bluez.Characteristic dbus object.
A corresponding remove_match for this object needs to be present during
tearing down of the event loop.
Without this the max match rules for a connection exceeds causing Android
framework to restart.

Change-Id: I26ce40d94e921346cf059a85dbff5a1f63ccb231
DRT:OMAPS00272452
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>